+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PR #39 Add cython headers to install, python / cmake packaging cleanup
 - PR #41 Python and Cython style cleanup, pre-commit hook
 - PR #44 Update all demos with Python API
+- PR #45 Improve documentation in haversine and point in polygon
 
 ## Bug Fixes